chore: update Book api
